fix(stream-engine): #45 ADR-0048 — canonical A·P·Q finalize for the day grain + hour/shift ::float hardening
The deployed day finalize (dayRollupSQL) kept a top-down oee = net/ideal_production
and BACK-SOLVED oee_p = oee/(oee_a·oee_q); when oee_p clamps to [0,1] the stored
oee no longer equals oee_a·oee_p·oee_q. So every NEW day row could re-break the
identity — the #885 historical reconcile only fixed existing rows. This closes the
code gap so day matches hour/shift/week/month.

- day.go: add dayOeeReconcileSQL (canonical factors-first: oee_a=running::float/
  available, oee_q=net/gross, oee_p=gross·available/(ideal_production·running) —
  day has no ideal_speed column, same algebra as week/month — each clamped [0,1],
  oee = product as the LAST step). Gated on OeeCanonicalAPQ exactly like grains/
  hour/shift, running right after the rollup; the legacy back-solve stays as the
  un-gated else-branch for parity. RunDay now takes `ca CountersAvail` (threaded
  from LoopGrains). ::float defeats integer division on the integer day columns;
  net=0 → oee_q=net/NULLIF(gross,0)=0 so zero-output days read oee=0, never 1.0.

- availability.go: add the ::float cast to hourOeeReconcileSQL/shiftOeeReconcileSQL
  Availability (running_time::float/available_time) — the latent trap flagged in
  #883/#885. Confirmed no-op on current live values (a_new == a_old: 1489/1495 hour,
  385/387 shift rows match stored to 1e-4; the few diffs are counters-avail/line-lead
  rows), while 368 hour rows would have been wrong under cast-less integer division.

- Tests: TestDayCanonicalReconcile + TestHourShiftReconcileFloatCast (string
  guards: day computes oee as the product of 3 bounded factors not back-solved,
  ::float present in day+hour+shift). New golden TestGoldenDayOeeReconcile proves
  the identity on real Postgres with prod-mirroring INTEGER running/available
  columns (cast regression → A=0 fails) and net=0 → oee=0 (no spurious 1.0).

go build / vet / test ./internal/rollup and -tags golden all green. No live DB
changes (history already reconciled by #885); forward-looking code fix only.

FOLLOW-UP addressed: this is the day-grain durability gap flagged in #885.

+++ services/stream-engine/internal/rollup/availability.go
@@ -367,12 +367,19 @@ const hourAvailFloorSQL = `
 // mirroring rollup/oee.go CanonicalOEE exactly: Performance is computed DIRECTLY
 // (gross/(ideal_speed·running/60)), so oee == oee_a·oee_p·oee_q by construction.
 // Replaces the legacy oee-p residual pass when engagedCanonical().
+//
+// ::float on Availability: running_time/available_time are INTEGER columns, so a
+// cast-less running_time/available_time would be integer division (e.g. 1800/3600
+// → 0). Today this is masked — the events pass writes a float oee_a from
+// extract(epoch…) before this reconcile runs, and on the rows this touches the
+// cast reproduces that same value (a_new == a_old, verified live) — but the cast
+// removes the integer-division trap for any future refactor (matches day/week/month).
 const shiftOeeReconcileSQL = `
 	UPDATE %[1]s.equipment_runtime_shift e SET
-	       oee_a = GREATEST(LEAST(COALESCE(e.running_time / NULLIF(e.available_time, 0), 0), 1), 0),
+	       oee_a = GREATEST(LEAST(COALESCE(e.running_time::float / NULLIF(e.available_time, 0), 0), 1), 0),
 	       oee_q = GREATEST(LEAST(COALESCE(e.net / NULLIF(e.gross, 0), 0), 1), 0),
 	       oee_p = GREATEST(LEAST(COALESCE(e.gross / NULLIF(e.ideal_speed * e.running_time / 60.0, 0), 0), 1), 0),
-	       oee   = GREATEST(LEAST(COALESCE(e.running_time / NULLIF(e.available_time, 0), 0), 1), 0)
+	       oee   = GREATEST(LEAST(COALESCE(e.running_time::float / NULLIF(e.available_time, 0), 0), 1), 0)
 	             * GREATEST(LEAST(COALESCE(e.gross / NULLIF(e.ideal_speed * e.running_time / 60.0, 0), 0), 1), 0)
 	             * GREATEST(LEAST(COALESCE(e.net / NULLIF(e.gross, 0), 0), 1), 0)
 	  FROM shift_elig el
@@ -381,10 +388,10 @@ const shiftOeeReconcileSQL = `
 
 const hourOeeReconcileSQL = `
 	UPDATE %[1]s.equipment_runtime_1hour e SET
-	       oee_a = GREATEST(LEAST(COALESCE(e.running_time / NULLIF(e.available_time, 0), 0), 1), 0),
+	       oee_a = GREATEST(LEAST(COALESCE(e.running_time::float / NULLIF(e.available_time, 0), 0), 1), 0),
 	       oee_q = GREATEST(LEAST(COALESCE(e.net / NULLIF(e.gross, 0), 0), 1), 0),
 	       oee_p = GREATEST(LEAST(COALESCE(e.gross / NULLIF(e.ideal_speed * e.running_time / 60.0, 0), 0), 1), 0),
-	       oee   = GREATEST(LEAST(COALESCE(e.running_time / NULLIF(e.available_time, 0), 0), 1), 0)
+	       oee   = GREATEST(LEAST(COALESCE(e.running_time::float / NULLIF(e.available_time, 0), 0), 1), 0)
 	             * GREATEST(LEAST(COALESCE(e.gross / NULLIF(e.ideal_speed * e.running_time / 60.0, 0), 0), 1), 0)
 	             * GREATEST(LEAST(COALESCE(e.net / NULLIF(e.gross, 0), 0), 1), 0)
 	  FROM hour_elig el
